refact: move page header to components root
